refactor:apply code review for validate password
