+++ src/main/java/orangetaxiteam/cocoman/domain/DefaultPasswordValidator.java
@@ -2,17 +2,16 @@ package orangetaxiteam.cocoman.domain;
 
 import orangetaxiteam.cocoman.domain.exceptions.BadRequestException;
 import orangetaxiteam.cocoman.domain.exceptions.ErrorCode;
-import org.springframework.beans.factory.annotation.Autowired;
+import orangetaxiteam.cocoman.infra.DefaultPasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Component
 public class DefaultPasswordValidator implements PasswordValidator {
-    @Autowired
-    private PasswordEncoder passwordEncoder;
+    private DefaultPasswordEncoder defaultPasswordEncoder = new DefaultPasswordEncoder();
 
     @Override
     public void validate(User user, String password) {
-        if (!passwordEncoder.matches(password,user.getPassword())) {
+        if (!defaultPasswordEncoder.matches(password,user.getPassword())) {
             throw new BadRequestException(
                     ErrorCode.SIGNIN_PASSWORD_DOES_NOT_MATCH,
                     "Password does not match");
