+++ src/main/java/orangetaxiteam/cocoman/domain/PasswordEncoder.java
@@ -1,7 +1,7 @@
 package orangetaxiteam.cocoman.domain;
 
 public interface PasswordEncoder {
-    public String encode(CharSequence var1);
+    public String encode(String var1);
 
-    public boolean matches(CharSequence var1, String var2);
+    public boolean matches(String var1, String var2);
 }
