+++ src/main/java/orangetaxiteam/cocoman/infra/DefaultPasswordEncoder.java
@@ -1,20 +1,20 @@
-package orangetaxiteam.cocoman.domain;
+package orangetaxiteam.cocoman.infra;
 
-import org.springframework.security.crypto.bcrypt.BCrypt;
+import orangetaxiteam.cocoman.domain.PasswordEncoder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DefaultPasswordEncoder implements PasswordEncoder{
+public class DefaultPasswordEncoder implements PasswordEncoder {
     private BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
 
     @Override
-    public String encode(CharSequence var1) {
+    public String encode(String var1) {
         return bCryptPasswordEncoder.encode(var1);
     }
 
     @Override
-    public boolean matches(CharSequence var1, String var2) {
+    public boolean matches(String var1, String var2) {
         if (bCryptPasswordEncoder.matches(var1,var2)) return true;
         else return false;
     }
